// src/services/types/platform/integrations.ts
export type IntegrationsApiResponse = {
  _id: string
  __v: number
  integrations: {
    section1: {
      title1_en: string
      title1_fa: string
      title2_en: string
      title2_fa: string
      text1_en: string
      text1_fa: string
      button1_en: string
      button1_fa: string
    }
    section2: {
      title1_en: string
      title1_fa: string
      text1_en: string
      text1_fa: string
      image1: string
    }
    section3: {
      title1_en: string
      title1_fa: string
      text1_en: string
      text1_fa: string
      button1_en: string
      button1_fa: string
      image1: string
    }
    section4: {
      title1_en: string
      title1_fa: string
      image1: string
      image2: string
      image3: string
      image4: string
      image5: string
      image6: string
      image7: string
      image8: string
      image9: string
      image10: string
      image11: string
      image12: string
      image13: string
      image14: string
      image15: string
      image16: string
      image17: string
      image18: string
      image19: string
      image20: string
      image21: string
      image22: string
      image23: string
      image24: string
      image25: string
    }
    section5: {
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      button1_en: string
      button1_fa: string
    }
  }
}